Description:     Normalize against trunk, in preparation for eventual merge to trunk.
Tested on:
    MacOSX/64 10.10.3 (amazon) w/serial & parallel
    (h5committest not required on this branch)

+++ features/metadata_cache_merge/src/H5AC.c
@@ -772,12 +772,11 @@ H5AC_flush(H5F_t *f, hid_t dxpl_id)
     /* For the flush, only the flags are really necessary in the trace file.
      * Write the result to catch occult errors.
      */
-    if ( ( f != NULL ) &&
-         ( f->shared != NULL ) &&
-         ( f->shared->cache != NULL ) &&
-         ( H5C_get_trace_file_ptr(f->shared->cache, &trace_file_ptr) >= 0 ) &&
-         ( trace_file_ptr != NULL ) )
-
+    if((f != NULL) &&
+            (f->shared != NULL) &&
+            (f->shared->cache != NULL) &&
+            (H5C_get_trace_file_ptr(f->shared->cache, &trace_file_ptr) >= 0) &&
+            (trace_file_ptr != NULL))
 	sprintf(trace, "H5AC_flush");
 #endif /* H5AC__TRACE_FILE_ENABLED */
 
@@ -1001,8 +1000,7 @@ H5AC_mark_entry_dirty(void *thing)
      * is really necessary in the trace file.  Write the result to catch
      * occult errors.
      */
-    if((H5C_get_trace_file_ptr_from_entry((const H5C_cache_entry_t *) thing, 
-                                          &trace_file_ptr) >= 0) &&
+    if((H5C_get_trace_file_ptr_from_entry((const H5C_cache_entry_t *) thing, &trace_file_ptr) >= 0) &&
             (NULL != trace_file_ptr))
         sprintf(trace, "%s 0x%lx", FUNC,
 	        (unsigned long)(((H5C_cache_entry_t *)thing)->addr));
@@ -1150,8 +1148,7 @@ H5AC_pin_protected_entry(void *thing)
     /* For the pin protected entry call, only the addr is really necessary
      * in the trace file.  Also write the result to catch occult errors.
      */
-    if((H5C_get_trace_file_ptr_from_entry((const H5C_cache_entry_t *)thing, 
-                                          &trace_file_ptr) >= 0) &&
+    if((H5C_get_trace_file_ptr_from_entry((const H5C_cache_entry_t *)thing, &trace_file_ptr) >= 0) &&
             (NULL != trace_file_ptr))
         sprintf(trace, "%s 0x%lx", FUNC,
 	        (unsigned long)(((H5C_cache_entry_t *)thing)->addr));
@@ -1199,8 +1196,7 @@ H5AC_create_flush_dependency(void * parent_thing, void * child_thing)
     HDassert(child_thing);
 
 #if H5AC__TRACE_FILE_ENABLED
-    if((H5C_get_trace_file_ptr_from_entry((H5C_cache_entry_t *)parent_thing, 
-                                          &trace_file_ptr) >= 0) &&
+    if((H5C_get_trace_file_ptr_from_entry((H5C_cache_entry_t *)parent_thing, &trace_file_ptr) >= 0) &&
             (NULL != trace_file_ptr))
         sprintf(trace, "%s %lx %lx",
                 FUNC,
@@ -1290,8 +1286,7 @@ H5AC_protect(H5F_t *f,
 
 
     /* Check for invalid access request */
-    if((0 == (H5F_INTENT(f) & H5F_ACC_RDWR)) && 
-       (0 == (flags & H5C__READ_ONLY_FLAG)))
+    if((0 == (H5F_INTENT(f) & H5F_ACC_RDWR)) &&  (0 == (flags & H5C__READ_ONLY_FLAG)))
 	HGOTO_ERROR(H5E_CACHE, H5E_BADVALUE, NULL, "no write intent on file")
 
 #if H5AC__TRACE_FILE_ENABLED
@@ -1311,7 +1306,6 @@ H5AC_protect(H5F_t *f,
 		flags);
     }
 #endif /* H5AC__TRACE_FILE_ENABLED */
-
 
     if ( flags & H5C__READ_ONLY_FLAG ) 
 	protect_flags |= H5C__READ_ONLY_FLAG;
@@ -1392,10 +1386,8 @@ H5AC_resize_entry(void *thing, size_t new_size)
      * really necessary in the trace file. Write the result to catch
      * occult errors.
      */
-    if ( ( H5C_get_trace_file_ptr_from_entry((H5C_cache_entry_t *)thing, 
-                                              &trace_file_ptr) >= 0 ) &&
-         ( NULL != trace_file_ptr ) )
-
+    if((H5C_get_trace_file_ptr_from_entry((H5C_cache_entry_t *)thing, &trace_file_ptr) >= 0) &&
+            (NULL != trace_file_ptr))
         sprintf(trace, "%s 0x%lx %d", FUNC,
 	        (unsigned long)(((H5C_cache_entry_t *)thing)->addr),
 		(int)new_size);
@@ -1460,10 +1452,8 @@ H5AC_unpin_entry(void *thing)
     /* For the unpin entry call, only the addr is really necessary
      * in the trace file.  Also write the result to catch occult errors.
      */
-    if ( ( H5C_get_trace_file_ptr_from_entry((H5C_cache_entry_t *)thing, 
-                                             &trace_file_ptr) >= 0 ) &&
-         ( NULL != trace_file_ptr ) )
-
+    if((H5C_get_trace_file_ptr_from_entry((H5C_cache_entry_t *)thing, &trace_file_ptr) >= 0) &&
+            (NULL != trace_file_ptr))
         sprintf(trace, "%s 0x%lx", FUNC,
 	        (unsigned long)(((H5C_cache_entry_t *)thing)->addr));
 #endif /* H5AC__TRACE_FILE_ENABLED */
@@ -1509,10 +1499,8 @@ H5AC_destroy_flush_dependency(void * parent_thing, void * child_thing)
     HDassert(child_thing);
 
 #if H5AC__TRACE_FILE_ENABLED
-    if ( ( H5C_get_trace_file_ptr_from_entry((H5C_cache_entry_t *)parent_thing,
-                                             &trace_file_ptr) >= 0 ) &&
-          ( NULL != trace_file_ptr ) )
-
+    if((H5C_get_trace_file_ptr_from_entry((H5C_cache_entry_t *)parent_thing, &trace_file_ptr) >= 0) &&
+            (NULL != trace_file_ptr))
         sprintf(trace, "%s %llx %llx",
                 FUNC,
 	        (unsigned long long)(((H5C_cache_entry_t *)parent_thing)->addr),
@@ -2170,7 +2158,7 @@ done:
          ( trace_file_ptr != NULL ) ) {
 
 	HDfprintf(trace_file_ptr,
-                  "%s %d %d %d %d \"%s\" %d %d %d %f %d %d %ld %d %f %f %d %f %f %d %d %d %f %f %d %d %d %d %f %d %d %d\n",
+                  "%s %d %d %d %d \"%s\" %d %d %d %f %d %d %ld %d %f %f %d %f %f %d %d %d %f %f %d %d %d %d %f %zu %d %d\n",
 		  "H5AC_set_cache_auto_resize_config",
 		  trace_config.version,
 		  (int)(trace_config.rpt_fcn_enabled),
@@ -3507,7 +3495,7 @@ H5AC_log_dirtied_entry(const H5AC_info_t * entry_ptr,
             }
 
             aux_ptr->d_slist_len += 1;
-            aux_ptr->dirty_bytes += (int32_t)entry_ptr->size;
+            aux_ptr->dirty_bytes += entry_ptr->size;
 #if H5AC_DEBUG_DIRTY_BYTES_CREATION
 	    aux_ptr->unprotect_dirty_bytes += entry_ptr->size;
 	    aux_ptr->unprotect_dirty_bytes_updates += 1;
@@ -3535,7 +3523,7 @@ H5AC_log_dirtied_entry(const H5AC_info_t * entry_ptr,
         } /* end if */
     } else {
 
-        aux_ptr->dirty_bytes += (int32_t)entry_ptr->size;
+        aux_ptr->dirty_bytes += entry_ptr->size;
 #if H5AC_DEBUG_DIRTY_BYTES_CREATION
         aux_ptr->unprotect_dirty_bytes += entry_size;
         aux_ptr->unprotect_dirty_bytes_updates += 1;
@@ -3748,7 +3736,7 @@ H5AC_log_inserted_entry(H5AC_t * cache_ptr,
             HGOTO_ERROR(H5E_CACHE, H5E_SYSTEM, FAIL, "Inserted entry in clean slist.")
     } /* end if */
 
-    aux_ptr->dirty_bytes += (int32_t)entry_ptr->size;
+    aux_ptr->dirty_bytes += entry_ptr->size;
 
 #if H5AC_DEBUG_DIRTY_BYTES_CREATION
     aux_ptr->insert_dirty_bytes += size;
@@ -3922,7 +3910,7 @@ H5AC_log_moved_entry(const H5F_t *f,
             slist_entry_ptr->magic = H5AC__H5AC_SLIST_ENTRY_T_MAGIC;
             slist_entry_ptr->addr  = new_addr;
 
-            aux_ptr->dirty_bytes += (int32_t)entry_size;
+            aux_ptr->dirty_bytes += entry_size;
 
 #if H5AC_DEBUG_DIRTY_BYTES_CREATION
             aux_ptr->move_dirty_bytes += entry_size;
@@ -3949,7 +3937,7 @@ H5AC_log_moved_entry(const H5F_t *f,
 
     } else if ( ! entry_dirty ) {
 
-        aux_ptr->dirty_bytes += (int32_t)entry_size;
+        aux_ptr->dirty_bytes += entry_size;
 
 #if H5AC_DEBUG_DIRTY_BYTES_CREATION
         aux_ptr->move_dirty_bytes += entry_size;
@@ -5016,15 +5004,15 @@ H5AC_run_sync_point(H5F_t *f,
 
 #if H5AC_DEBUG_DIRTY_BYTES_CREATION
     HDfprintf(stdout,
-              "%d:H5AC_propagate...:%d: (u/uu/i/iu/r/ru) = %d/%d/%d/%d/%d/%d\n",
-              (int)(aux_ptr->mpi_rank),
-              (int)(aux_ptr->dirty_bytes_propagations),
-              (int)(aux_ptr->unprotect_dirty_bytes),
-              (int)(aux_ptr->unprotect_dirty_bytes_updates),
-              (int)(aux_ptr->insert_dirty_bytes),
-              (int)(aux_ptr->insert_dirty_bytes_updates),
-              (int)(aux_ptr->rename_dirty_bytes),
-              (int)(aux_ptr->rename_dirty_bytes_updates));
+              "%d:H5AC_propagate...:%u: (u/uu/i/iu/r/ru) = %zu/%u/%zu/%u/%zu/%u\n",
+              aux_ptr->mpi_rank,
+              aux_ptr->dirty_bytes_propagations,
+              aux_ptr->unprotect_dirty_bytes,
+              aux_ptr->unprotect_dirty_bytes_updates,
+              aux_ptr->insert_dirty_bytes,
+              aux_ptr->insert_dirty_bytes_updates,
+              aux_ptr->rename_dirty_bytes,
+              aux_ptr->rename_dirty_bytes_updates);
 #endif /* H5AC_DEBUG_DIRTY_BYTES_CREATION */
 
     switch(aux_ptr->metadata_write_strategy) {
